New major release include: -solution enumeration cf. favier&al cp 2009 -different bug correction on vac and btd -command line cleaning -short user doc added in doc dir.

+++ toulbar2/src/tb2variable.hpp
@@ -77,7 +77,7 @@ public:
     ConstraintList *getConstrs() {return &constrs;}
     int getDegree() {return constrs.getSize();}
     int getTrueDegree();
-    double getWeightedDegree();
+    Long getWeightedDegree();
     DLink<ConstraintLink> *link(Constraint *c, int index);
     void sortConstraints();
     virtual void eliminate() {cout << "variable elimination not implemented!" << endl;};
